Some uint/size_t compile fixes
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@6114 3db68df2-c116-0410-a063-a993310a9797

+++ source/network/NetLog.cpp
@@ -154,7 +154,7 @@ void CNetLogSink::DoBulkSink( const CNetLogEvent* pEvents, size_t eventCount )
 	if ( m_Closed ) return;
 
 	// Allocate new array which will store the events that will be logged
-	pIndices = new unsigned[ eventCount ];
+	pIndices = new size_t[ eventCount ];
 	if ( !pIndices ) 
 	{
 		throw std::bad_alloc();
